:bug: fix spk file api #114

+++ modules/core/models/tts/ChatTtsModel.py
@@ -1,6 +1,7 @@
 from typing import Any, Generator, Union
 
 import numpy as np
+import torch
 
 from modules.core.models.tts.InerCache import InferCache
 from modules.core.models.TTSModel import TTSModel
@@ -22,7 +23,7 @@ class ChatTTSModel(TTSModel):
     def create_speaker_from_seed(seed: int):
         chat = load_chat_tts()
         with SeedContext(seed):
-            token = chat._sample_random_speaker()
+            token = chat._sample_random_speaker().float()
             spk = TTSSpeaker.empty()
             spk.set_token(tokens=[token], model_id=ChatTTSModel.model_id)
             spk.set_name(f"spk[seed={seed}]")
